changed package names, upgraded some plugins
