+++ exercises/mvn.05.exercise-multi/simple-weather/src/test/java/course/maven/weather/VisualCrossingRetrieverTest.java
@@ -1,0 +1,26 @@
+package course.maven.weather;
+
+import static org.junit.Assert.assertThat;
+
+import java.io.InputStream;
+
+import org.apache.commons.io.IOUtils;
+import org.hamcrest.CoreMatchers;
+import org.junit.Test;
+
+import course.maven.weather.VisualCrossingRetriever;
+
+public class VisualCrossingRetrieverTest {
+
+	@Test
+	public void simpleTest() throws Exception {
+		VisualCrossingRetriever retriever = new VisualCrossingRetriever();
+		final String content;
+		try (InputStream is = retriever.retrieve("köln")) {
+			content = IOUtils.toString(is);	
+		}
+
+		assertThat(content, CoreMatchers.containsString("Köln, Nordrhein-Westfalen, Deutschland"));
+	}
+
+}
