+++ solutions/mvn.05.solution-multi/simple-weather/src/main/java/course/maven/weather/ProxyAuthenticator.java
@@ -1,0 +1,17 @@
+package course.maven.weather;
+
+import java.net.Authenticator;
+import java.net.PasswordAuthentication;
+
+public final class ProxyAuthenticator extends Authenticator {
+	private String user, password;  
+	  
+    public ProxyAuthenticator(String user, String password) {  
+        this.user = user;  
+        this.password = password;  
+    }  
+  
+    protected PasswordAuthentication getPasswordAuthentication() {  
+        return new PasswordAuthentication(user, password.toCharArray());  
+    }  
+}
