--- solutions/mvn.05.solution-multi/simple-webapp/src/main/java/com/sonatype/maven/web/SimpleServlet.java
@@ -1,21 +1,0 @@
-package com.sonatype.maven.web;
-
-import java.io.IOException;
-import java.io.PrintWriter;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@WebServlet("/simple")
-public class SimpleServlet extends HttpServlet {
-	@Override
-	public void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		PrintWriter out = response.getWriter();
-		out.println("SimpleServlet Executed");
-		out.flush();
-		out.close();
-	}
-}
